refactor(headless-styles): combine select option api into select api files
