+++ packages/headless-styles/src/components/Select/selectCSS.module.css
@@ -18,6 +18,11 @@
   right: 0.5rem;
 }
 
+.selectOption {
+  background-color: initial;
+  color: initial;
+}
+
 /* sizes */
 
 .mSelectBase {
