+++ packages/headless-styles/src/components/Select/selectCSS.ts
@@ -2,9 +2,11 @@ import { createClassNameProp } from '../../utils/helpers'
 import {
   createSelectClasses,
   createSelectProps,
+  createSelectOptionProps,
   getDefaultSelectOptions,
+  getDefaultSelectOptionOptions,
 } from './shared'
-import type { SelectOptions } from './types'
+import type { SelectOptions, SelectOptionOptions } from './types'
 import styles from './selectCSS.module.css'
 
 const SELECT = 'ps-select'
@@ -34,5 +36,15 @@ export function getSelectProps(options?: SelectOptions) {
       ...props.iconWrapper,
       ...createClassNameProp(`${SELECT}-icon ${styles.selectIcon}`),
     },
+  }
+}
+
+export function getSelectOptionProps(options?: SelectOptionOptions) {
+  const defaultOptions = getDefaultSelectOptionOptions(options)
+  const props = createSelectOptionProps(defaultOptions)
+
+  return {
+    ...props,
+    ...createClassNameProp(`${SELECT}-option ${styles.selectOption}`),
   }
 }
