--- packages/headless-styles/src/components/Select/selectOptionCSS.module.css
@@ -1,4 +1,0 @@
-.selectOption {
-  background-color: initial;
-  color: initial;
-}
