+++ packages/headless-styles/src/index.ts
@@ -61,11 +61,14 @@ export { getJSProgressProps } from './components/Progress/progressJS'
 export { getRadioProps } from './components/Radio/radioCSS'
 export { getJSRadioProps } from './components/Radio/radioJS'
 
-export { getSelectProps } from './components/Select/selectCSS'
-export { getJSSelectProps } from './components/Select/selectJS'
-
-export { getSelectOptionProps } from './components/Select/selectOptionCSS'
-export { getJSSelectOptionProps } from './components/Select/selectOptionJS'
+export {
+  getSelectProps,
+  getSelectOptionProps,
+} from './components/Select/selectCSS'
+export {
+  getJSSelectProps,
+  getJSSelectOptionProps,
+} from './components/Select/selectJS'
 
 export { getSkeletonProps } from './components/Skeleton/skeletonCSS'
 export { getJSSkeletonProps } from './components/Skeleton/skeletonJS'
